- add Delete Account

+++ app/Http/Controllers/Api/V1/UserController.php
@@ -243,20 +243,14 @@ class UserController extends Controller
      */
     public function delete(Request $request)
     {
-//        $validator = Validator::make($request->all(), [
-//            //'password' => ['required', 'string'],
-//        ]);
-//
-//        if ($validator->fails()) {
-//            return $this->errorResponse($validator->errors()->first(), 422);
-//        }
+
 
         $user = auth()->user();
 
-//        if (!Hash::check($request->password, $user->password)) {
-//            return $this->errorResponse(trans('messages.incorrect_password'), 422);
-//        }
 
+        if (!$user) {
+            return $this->errorResponse(trans('messages.login_first'), 422);
+        }
         $user->delete();
 
         return $this->successResponse(trans('messages.account_deleted_successfully'));
@@ -296,7 +290,6 @@ class UserController extends Controller
 
         return $this->successResponse('FCM updated successfully');
     }
-
 
     /**
      * Prepare user data for response.
